Add PDF + Send Mail + change some entity + Add discount coupon + Strip payment

+++ src/main/java/tn/esprit/spring/entities/Product.java
@@ -5,14 +5,14 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
+
 
 @Entity
 public class Product {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String label;
 	private float price;
 	private int quantity;
@@ -21,13 +21,13 @@ public class Product {
 	@Column(unique=true)
 	private String barCode;
 	private String image;
+	
 
-
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
